Make server sim state deep-copyable
We need to save independent copies of this state in order to do lag
compensation. Right now, we have a pseudo-deep copy function, but
technically any modification to the nested objects of the resulting copy
would still affect the original. Very unpleasant! Let's fix.

Implementation note: We add a utility function to deep copy an
array of items. Godot doesn't have generic method signatures, so
to keep it simple, we just return a non-parametrized Array Type.
Unfortunately, we're trying to assing this return type to parametrized
Array-type variables, which makes a type error. The workaround is to
use the Array.assign method to assign the non-parametrized copy to
the parametrized var