Remove lossy compression for own-character state
The server sends each client the state of its own character, which is
used for reconciliation. Right now, we're compressing position and
velocity (4 bytes per float -> 2 bytes), which causes desyncs in the
reconciliation because of floating point errors. It's probably not worth
the bandwidth savings, seeing as this state is a flat cost per packet
send to each client (rather than scaling with the # of players).

Let's just serialize the full information.