Keep displaying remote state when dead
Problem: When the player dies, right now everything on the screen stops,
and the server doesn't send any state updates until respawn. This does
not feel very responsive for the player. In most online games, you'll
see the world continuing to update when you die, so this is also bad
because it's unfamiliar behavior. Finally, we might want to add
spectator mode when dead.

To fix, let's keep sending state to the player even when they're dead.
The main challenge is handling the starting/stopping of the client's
own state prediction and displaying. AKA, movement and sounds and UI
etc. for their own character. For now, we're taking the dead simple
approach of manually doing a bunch of cleanup when we get the spawn
or death RPCs from the server, as well as some not-so-clean flow control
which hides UI elements when dead. Definitely should be improved in
the future, but this is a good first step.

Another major win from this commit is separating the remote display
logic and own state prediction logic into functions. Will make it easier
to handle death/respawn cleanly in the future.