Add feedback on remote character death
In our playtest, the one player said it's unclear when you've actually
killed an enemy player vs. when they disconnect. In both cases, the
model just disappears instantly. Both for clarity and to add juice,
let's make it possible to play an animation when another player's
character dies.

We have it so that when a remote character dies, the client deletes all
its associated objects, including playermodel. As a result, we can't
just call an animation function on the remote character object when
it dies. Why not keep the objects around upon death? It's nice to
be able to have 1-to-1 correspondence between entities in the state
and existent objects. And we don't keep the state around because it
eliminates the risk that dead players would keep interacting with the
world.

Instead, we use the approach of making a separate, short-lived object
that just plays the death animation and then frees itself. This may seem
excessive / duplication-y for just the simple fade animation we have
now, but this separation gives us more flexibility in the future for
different death effects like ragdolls, gibbing, etc. which have little
in common with the basic remote character model.