Use same ability trigger class for RPCs and client
Right now, we trigger remote character abilities on client by sending
all the bits of data we need as separate parameters to the RPC.

On the client side, we package all these bits these up as a
RemoteCharacterAbilityTrigger object, which is useful because we
want to defer when we display the trigger until we are displaying
an up-to-date view of the remote state.

It is a bit annoying to have to change the function signature for the
RPC, its public wrapper function, the signal we emit when the RPC
is received on client, and the callback function to that signal whenever
we'd want to add/remove some data from the Trigger. It would be
simple to get rid of that extra wiring, but keeping the RPC's separate
is generally good design because it means the client class doesn't
need to know about the network implementation details at all.
Instead, we can solve the problem by passing the same data class,
RemoteCharacterAbilityTrigger, through all these methods.