Detect enemy death via state change, not RPC
Right now, the server RPC's the client to let it know when a remote
player has died, so it can play a death animation. The problem here
is the client renders the world it knows about from state snapshots,
which can be out of sync with RPCs.

Specifically, we delete enemy character 3D models when their entity
is no longer in the latest remote state. However, the RPC can arrive
much later due to network problems. So the character will disappear,
and then their body will reappear a bit later to show the death
animation.

Let's fix this by using the same check for both object deletion
and playing the death animation.

In the future, we'll probably use RPC's to create and delete entities
from the client state; this is much more scalable when there's lots
of entities. But this is a fine short term solution.